Fix style issues

+++ src/game/graphicsobject.cpp
@@ -1,7 +1,7 @@
 #include "graphicsobject.h"
 
-GraphicsObject::GraphicsObject(QGraphicsView* graphics_view, qreal width,
-                               qreal height, qreal x, qreal y)
+GraphicsObject::GraphicsObject(QGraphicsView* graphics_view, float width,
+                               float height, float x, float y)
     : graphics_view_(graphics_view), width_(width), height_(height) {
   setPos(x, y);
 }
